==> nanotssqtl/1_getTSSbed.ipynb <==
import pandas as pd

input_file = '/mnt/hpc/home/xuxinran/REF/hg19/FANTOM5_CAGEpeak_TSS_human.bed.gz'
output_file = '/mnt/hpc/home/xuxinran/DirectSeq/nanotssqtl/TSSregion.bed'

df = pd.read_csv(input_file, skiprows=1, sep='\t', header=None, usecols=[0, 1, 2, 3, 4, 5], names=['chrom', 'start', 'end', 'name', 'score', 'strand'])
autosomes = [f'chr{i}' for i in range(1, 23)]
autosomal_df = df[df['chrom'].isin(autosomes)]
# 这里扩一下TSS的范围 |——50bp——CAGE peak——20bp——|
f_df = autosomal_df[autosomal_df['strand'] == '+']
f_df['start'] = f_df['start'] - 50
f_df['end'] = f_df['end'] + 20
r_df = autosomal_df[autosomal_df['strand'] == '-']
r_df['start'] = r_df['start'] - 20
r_df['end'] = r_df['end'] + 50
autosomal_df = pd.concat([f_df, r_df])
autosomal_df = autosomal_df.drop_duplicates().sort_values(by=['chrom', 'start'])
autosomal_df.to_csv(output_file, sep='\t', header=False, index=False)
