==> nanotssqtl/4_fdr.ipynb <==
import numpy as np
import os
import glob
import pandas as pd
from statsmodels.sandbox.stats.multicomp import multipletests
import numpy as np
from subprocess import call

dir_pre = "/mnt/hpc/home/xuxinran/DirectSeq/data/zhaolin_240206/240201-zhaolin-RNA-merge/v0.7.2/tss/nano_merge"

output_path = f"{dir_pre}_signal_tssqtl.csv"
output_path_all = f"{dir_pre}_tss_SNP.csv"

files = glob.glob(f'{dir_pre}_haplotype_chr*_tmp.csv')
dfs = [pd.read_csv(file) for file in files]
df = pd.concat(dfs, ignore_index=True)
df = df.sort_values(by=['chrom', 'snp_pos_1base'])
df = df.reset_index(drop=True)

p_adjusted = multipletests(df["p_value"], method='fdr_bh')
df['fdr'] = p_adjusted[1]
print("pvalue小于0.05的个数：", len(df[df['p_value']<0.05]))
print("pvalue小于0.01的个数：", len(df[df['p_value']<0.01]))
print("fdr小于0.1的个数：", len(df[df['fdr']<0.1]))

result_df = df[df['p_value']<0.05]
result_df.to_csv(output_path, index=False)
df.to_csv(output_path_all, index=False)

call(f"rm {dir_pre}_haplotype_chr*_tmp.csv {dir_pre}_overlap.bed", shell=True)
